requirements.txt
google-adk
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
# statsmodels>=0.14.0  # Optional - uncomment if you want detailed statistical analysis 